+++ resources/views/components/dashboard/events.blade.php
@@ -4,10 +4,14 @@
     $hasAttendancePollEditPermission = \Illuminate\Support\Facades\Auth::user()?->hasPermission(\App\Models\AttendancePoll::ATTENDANCE_POLL_EDIT_PERMISSION) ?? false;
 
 @endphp
-@if($hasShowPermission)
-    <div class="flex-1 flex flex-col mb-3">
+@if($hasShowPermission || $hasAttendancePollEditPermission)
+    <div class="flex-1 flex flex-col mb-3 bg-white shadow-sm sm:rounded-lg p-4 text-center">
+        <div class="flex justify-center items-center gap-2 text-xl">
+            <i class="fa-solid fa-calendar"></i>
+            <span>{{__('event management')}}</span>
+        </div>
+        <hr class="my-3">
         @php($pollList = \App\Models\AttendancePoll::query()->where('closing_at','>',now()))
-        <div class="bg-white shadow-sm sm:rounded-lg p-4 text-center">
         @if($pollList->count())
             <header class="font-bold mb-3">{{__('Active polls')}}</header>
             <ul class="grid text-sm gap-2 text-white">
@@ -36,9 +40,15 @@
                     </li>
                 @endforeach
             </ul>
-            @else
-                <span class="text-gray-500">{{__('no active polls')}}</span>
-            @endif
-        </div>
+        @else
+            <span class="text-gray-500">{{__('no active polls')}}</span>
+        @endif
+        @if($futureDisabledCnt = \App\Models\Event::query()->where('start', '>', now())->where('enabled', false)->count())
+            <hr class="my-3">
+            <div class="text-yellow-500">
+                <i class="fa-solid fa-ban"></i>
+                <span>{{__(':cnt future but disabled event(s)', ['cnt' => $futureDisabledCnt])}}</span>
+            </div>
+        @endif
     </div>
 @endif