+++ resources/views/components/dashboard/members.blade.php
@@ -1,14 +1,18 @@
 @php
     /** @var \App\Models\Member $member */
-    $hasShowPermission = \Illuminate\Support\Facades\Auth::user()?->hasPermission(\App\Models\Member::MEMBER_SHOW_PERMISSION) ?? false;
     $hasEditPermission = \Illuminate\Support\Facades\Auth::user()?->hasPermission(\App\Models\Member::MEMBER_EDIT_PERMISSION) ?? false;
 @endphp
-@if($hasShowPermission || $hasEditPermission)
-    <div class="flex-1 flex flex-col mb-3">
+@if($hasEditPermission)
+    <div class="flex-1 flex flex-col mb-3 bg-white shadow-sm sm:rounded-lg p-4 text-center">
+        <div class="flex justify-center items-center gap-2 text-xl">
+            <i class="fa-solid fa-user-pen"></i>
+            <span>{{__('member management')}}</span>
+        </div>
+        <hr class="my-3">
         @php($membersWithoutBirthday = \App\Models\Member::getAllFiltered()->whereNull('birthday')->get())
-        <div class="bg-white shadow-sm sm:rounded-lg p-4 text-center">
-            @if($membersWithoutBirthday->isNotEmpty())
-            <header class="font-bold mb-3">{{__('Member without birthdays')}}</header>
+        @if($membersWithoutBirthday->isNotEmpty())
+            @php($isDataConsistent = false)
+            <header class="font-bold mb-3">{{__('without birthdays')}}</header>
             <ul class="grid text-sm gap-2 text-white">
                 @foreach($membersWithoutBirthday as $member)
                     <li class="flex p-1 bg-red-600 rounded gap-4 justify-center">
@@ -21,9 +25,17 @@
                     </li>
                 @endforeach
             </ul>
-            @else
-                <span class="text-gray-500">{{__('member data is consistent')}}</span>
-            @endif
+        @endif
+        @if($missingEmailCnt = \App\Models\Member::getAllFiltered()->whereNull('email')->count())
+        @php($isDataConsistent = false)
+        <hr class="my-3">
+        <div class="text-red-800">
+            <i class="fa-solid fa-at"></i>
+            <span>{{__(':cnt email(s) missing', ['cnt' => $missingEmailCnt])}}</span>
         </div>
+        @endif
+        @if($isDataConsistent)
+            <span class="text-green-800"><i class="fa-solid fa-user-check"></i>{{__('member data is consistent')}}</span>
+        @endif
     </div>
 @endif