+++ resources/views/components/dashboard/sponsoring.blade.php
@@ -3,12 +3,16 @@
     $hasShowPermission = \Illuminate\Support\Facades\Auth::user()?->hasPermission(\App\Models\Sponsoring\Contract::SPONSORING_SHOW_PERMISSION) ?? false;
     $hasEditPermission = \Illuminate\Support\Facades\Auth::user()?->hasPermission(\App\Models\Sponsoring\Contract::SPONSORING_EDIT_PERMISSION) ?? false;
 @endphp
-@if($hasShowPermission)
-    <div class="flex-1 flex flex-col mb-3">
-        @php($periodList = \App\Models\Sponsoring\Period::query()->where('end','>',now()))
-        <div class="bg-white shadow-sm sm:rounded-lg p-4 text-center">
+@if($hasShowPermission || $hasEditPermission)
+    @php($periodList = \App\Models\Sponsoring\Period::query()->where('end','>',now()))
+    <div class="flex-1 flex flex-col mb-3 bg-white shadow-sm sm:rounded-lg p-4 text-center">
+        <div class="flex justify-center items-center gap-2 text-xl">
+            <i class="fa-solid fa-file-contract"></i>
+            <span>{{__('sponsor management')}}</span>
+        </div>
+        <hr class="my-3">
         @if($periodList->count())
-            <header class="font-bold mb-3">{{__('Active and upcoming sponsor periods')}}</header>
+            <header class="font-bold mb-3">{{__('Active and upcoming periods')}}</header>
             <ul class="grid text-sm gap-2 text-white">
                 @foreach($periodList->get() as $period)
                     <li class="flex flex-wrap bg-cyan-700 p-1 rounded items-center justify-center gap-4">
@@ -33,9 +37,8 @@
                     </li>
                 @endforeach
             </ul>
-            @else
-                <span class="text-gray-500">{{__('no sponsor periods to show')}}</span>
-            @endif
-        </div>
+        @else
+            <span class="text-gray-500">{{__('no sponsor periods to show')}}</span>
+        @endif
     </div>
 @endif